move the scroll indicator to footer and fix it
